# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Documentation in Node.js Using Swagger",
        "description": "User Restful webservice application",
     "version": "1.0.0",
        "contact": {
          "name": "Sarvesh Upadhyay",
          "email": "sarvesh@web.com",
          "url": "http://localhost:2023/api-docs"
        }
    },
    "host": "localhost:2023/api",
    "basePath": "/",
...
    "produces": ["application/json"],
    "paths": {
...
      "/getById/{id}": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Id of the user which is to be searched",
            "type": "string"
          }
        ],
        "produces": ["application/json"],
        "get": {
          "tags": ["Users"],
          "summary": "Get user by id",
          "responses": {
            "200": {
              "description": "ok"
            },
            "404": {
              "description": "not found"
            }
          }
        }
      },
...
    }
  }